Improve handling of invalid MongoDB archive files
If you try to download a version of MongoDB that is not available
(eg. MongoDB 2.6.12 for Debian 7),
mongodb-download fails to properly handle the download error.
Instead, it creates an empty *.targz file.

Compounding on this problem, the `Decompress` library does
not properly handle the error, when it attemps to extract the
empty *.targz.

To help alleviate these issues, I'm using good ol' `tar`
(which properly handles the exception),
and I've include a very verbose debug log, to help anyone else
experiencing this issue.

A longer-term solution would be to fix mongodb-download,
so that it properly catches download errors.